backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PPT Processing
python-pptx==0.6.23
Pillow==10.2.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Utilities
python-magic==0.4.27
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
